Change metrics package to metric and rename the exported functions to more accurately represent their purpose.
