+++ internal/servers/controller/internal/metric/api_test.go
@@ -1,4 +1,4 @@
-package metrics
+package metric
 
 import (
 	"net/http"
